+++ FileHelpers/FileTransformEngine/FileTransformEngine.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Text;
+using FileHelpers.FileTransformEngine;
 using FileHelpers.Helpers;
 using FileHelpers.Streams;
